+++ src/api/init.ts
@@ -196,6 +196,32 @@ export async function startApiServer(datastore: DataStore, chainId: ChainID): Pr
     regexp: pathToRegex.pathToRegexp(endpoint.path),
   }));
 
+  // Manual route definitions for the /v2/ proxied endpoints
+  routes.push({
+    path: '/v2/pox',
+    regexp: /^\/v2\/pox(.*)/,
+  });
+  routes.push({
+    path: '/v2/info',
+    regexp: /^\/v2\/info(.*)/,
+  });
+  routes.push({
+    path: '/v2/accounts/*',
+    regexp: /^\/v2\/accounts(.*)/,
+  });
+  routes.push({
+    path: '/v2/contracts/call-read/*',
+    regexp: /^\/v2\/contracts\/call-read(.*)/,
+  });
+  routes.push({
+    path: '/v2/map_entry/*',
+    regexp: /^\/v2\/map_entry(.*)/,
+  });
+  routes.push({
+    path: '/v2/*',
+    regexp: /^\/v2(.*)/,
+  });
+
   const server = createServer(app);
 
   const serverSockets = new Set<Socket>();